refactor(specs): remove the primary-job constructors for bootstrap methods handled by the overlay
CreatePrimaryJobViaInitdb, CreatePrimaryJobViaRestoreSnapshot,
CreatePrimaryJobViaRecovery, CreatePrimaryJobViaPgBaseBackup,
JoinReplicaInstance and RestoreReplicaInstance built the Jobs used by
the instance subcommands removed in the previous commit, so they and
their dedicated jobRole values are dead code now that bootstrap goes
through the pod overlay. CreatePrimaryJob is renamed to CreateRoleJob
since it now only builds the generic role-tagged Job shell shared by
the remaining callers (major upgrade).
