+++ pkg/specs/bootstrap_overlay_test.go
@@ -48,110 +48,6 @@ var _ = Describe("ApplyBootstrapOverlay", func() {
 		}
 	}
 
-	// expectArgParity asserts that the overlay appends --bootstrap-mode plus
-	// exactly the same flags the equivalent bootstrap Job command carried after
-	// its "instance <subcommand>" prefix. The mode name is passed explicitly
-	// because it can differ from the old subcommand (initdb runs "instance init").
-	expectArgParity := func(mode string, jobCommand []string, instruction BootstrapInstruction) {
-		pod := runPod()
-		Expect(ApplyBootstrapOverlay(pod, instruction)).To(Succeed())
-
-		appended := pod.Spec.Containers[0].Command[3:]
-		Expect(appended[0]).To(Equal("--bootstrap-mode=" + mode))
-		Expect(appended[1:]).To(Equal(jobCommand[3:]))
-	}
-
-	It("matches the initdb Job command", func() {
-		cluster := apiv1.Cluster{
-			ObjectMeta: metav1.ObjectMeta{Name: "cluster"},
-			Spec: apiv1.ClusterSpec{
-				WalStorage: &apiv1.StorageConfiguration{Size: "1Gi"},
-				Bootstrap: &apiv1.BootstrapConfiguration{
-					InitDB: &apiv1.BootstrapInitDB{
-						Database:               "app",
-						Owner:                  "app",
-						Encoding:               "UTF8",
-						PostInitSQL:            []string{"SELECT 1"},
-						PostInitApplicationSQL: []string{"SELECT 2"},
-						PostInitTemplateSQL:    []string{"SELECT 3"},
-						PostInitApplicationSQLRefs: &apiv1.SQLRefs{
-							SecretRefs: []apiv1.SecretKeySelector{
-								{Key: "k", LocalObjectReference: apiv1.LocalObjectReference{Name: "s"}},
-							},
-						},
-					},
-				},
-			},
-		}
-		job := CreatePrimaryJobViaInitdb(cluster, 0)
-		expectArgParity("initdb", job.Spec.Template.Spec.Containers[0].Command, NewInitDBInstruction(cluster))
-	})
-
-	It("matches the join Job command", func() {
-		cluster := apiv1.Cluster{
-			ObjectMeta: metav1.ObjectMeta{Name: "cluster"},
-			Spec:       apiv1.ClusterSpec{WalStorage: &apiv1.StorageConfiguration{Size: "1Gi"}},
-		}
-		job := JoinReplicaInstance(cluster, 0)
-		expectArgParity("join", job.Spec.Template.Spec.Containers[0].Command, NewJoinInstruction(cluster))
-	})
-
-	It("matches the pgbasebackup Job command", func() {
-		cluster := apiv1.Cluster{
-			ObjectMeta: metav1.ObjectMeta{Name: "cluster"},
-			Spec:       apiv1.ClusterSpec{WalStorage: &apiv1.StorageConfiguration{Size: "1Gi"}},
-		}
-		job := CreatePrimaryJobViaPgBaseBackup(cluster, 0)
-		expectArgParity("pgbasebackup", job.Spec.Template.Spec.Containers[0].Command, NewPgBaseBackupInstruction(cluster))
-	})
-
-	It("matches the recovery (restore) Job command", func() {
-		cluster := apiv1.Cluster{
-			ObjectMeta: metav1.ObjectMeta{Name: "cluster"},
-			Spec: apiv1.ClusterSpec{
-				WalStorage: &apiv1.StorageConfiguration{Size: "1Gi"},
-				Bootstrap:  &apiv1.BootstrapConfiguration{Recovery: &apiv1.BootstrapRecovery{}},
-			},
-		}
-		job := CreatePrimaryJobViaRecovery(cluster, 0, nil)
-		expectArgParity("restore", job.Spec.Template.Spec.Containers[0].Command, NewRecoveryInstruction(cluster, nil))
-	})
-
-	It("matches the restoresnapshot primary Job command", func() {
-		cluster := apiv1.Cluster{
-			ObjectMeta: metav1.ObjectMeta{Name: "cluster"},
-			Spec: apiv1.ClusterSpec{
-				WalStorage: &apiv1.StorageConfiguration{Size: "1Gi"},
-				Bootstrap:  &apiv1.BootstrapConfiguration{Recovery: &apiv1.BootstrapRecovery{}},
-			},
-		}
-		object := &metav1.ObjectMeta{
-			Annotations: map[string]string{
-				utils.BackupLabelFileAnnotationName:         "bGFiZWw=",
-				utils.BackupTablespaceMapFileAnnotationName: "bWFw",
-			},
-		}
-		job := CreatePrimaryJobViaRestoreSnapshot(cluster, 0, object, nil)
-		expectArgParity(
-			"restoresnapshot",
-			job.Spec.Template.Spec.Containers[0].Command,
-			NewRestoreSnapshotInstruction(cluster, object, nil),
-		)
-	})
-
-	It("matches the restoresnapshot replica Job command", func() {
-		cluster := apiv1.Cluster{
-			ObjectMeta: metav1.ObjectMeta{Name: "cluster"},
-			Spec:       apiv1.ClusterSpec{WalStorage: &apiv1.StorageConfiguration{Size: "1Gi"}},
-		}
-		job := RestoreReplicaInstance(cluster, 0)
-		expectArgParity(
-			"restoresnapshot",
-			job.Spec.Template.Spec.Containers[0].Command,
-			NewRestoreSnapshotReplicaInstruction(cluster),
-		)
-	})
-
 	It("stamps the bootstrap annotation and preserves the drift annotations", func(ctx SpecContext) {
 		cluster := apiv1.Cluster{
 			ObjectMeta: metav1.ObjectMeta{Name: "test-cluster", Namespace: "default"},
